Add user dictionary.
I use hunspell -d instead of hunspell -p to specify the user dictionary
because I want to use flags (such as /M and /S) in the dictionary.

Flags I use so far:

/M for words that can have the possesive "'s" appended
/S for words that can have the plural "s" appended
/Y for words that can have the adverbial "ly" appended

The rules (in the hunspell-provided en_US.aff) are flexible enough to
deal with special cases. For example, /S knows about the rule "if the
stem ends with 'y', form the plural by replacing the 'y' by 'ies'".